requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
langchain
langchain-community
faiss-cpu
pypdf
python-dotenv
numpy
groq
